loop/requirements.txt
anthropic==0.49.0
openai==1.68.2
python-dotenv==1.0.1
